updated hook definitions added some more smtp testing added some debug output updated modules

+++ servers.h
@@ -14,8 +14,8 @@
 extern "C" {
 #endif
     
-ssize_t rumble_server_execute_hooks(sessionHandle* session, cvector* hooks, unsigned short flags);
-ssize_t rumble_server_schedule_hooks(masterHandle* handle, sessionHandle* session, unsigned short flags);
+ssize_t rumble_server_execute_hooks(sessionHandle* session, cvector* hooks, uint32_t flags);
+ssize_t rumble_server_schedule_hooks(masterHandle* handle, sessionHandle* session, uint32_t flags);
 
 void* rumble_smtp_init(void* m);
 void* rumble_pop3_init(void* m);
